refactor: Made all Option class plurals Renamed all the "option" class/method to plurals & moved them to package commons.options
