+++ src/main/java/edu/lorsenmarek/backend/commons/options/SerieSearchOptions.java
@@ -1,4 +1,4 @@
-package edu.lorsenmarek.backend.commons;
+package edu.lorsenmarek.backend.commons.options;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,6 +7,6 @@ import lombok.Data;
 @Builder
 @Data
 @AllArgsConstructor
-public class SerieSearchOption {
+public class SerieSearchOptions {
     private String title;
 }
