+++ src/main/java/edu/lorsenmarek/backend/repositories/SerieRepositoryCustom.java
@@ -1,10 +1,10 @@
 package edu.lorsenmarek.backend.repositories;
 
-import edu.lorsenmarek.backend.commons.SerieSearchOption;
+import edu.lorsenmarek.backend.commons.options.SerieSearchOptions;
 import edu.lorsenmarek.backend.models.Serie;
 
 import java.util.List;
 
 public interface SerieRepositoryCustom {
-    List<Serie> searchByOption(SerieSearchOption options);
+    List<Serie> searchByOptions(SerieSearchOptions options);
 }
